# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvidenceCaseResponse": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "horizon": {
            "title": "Horizon",
            "type": "string"
          },
          "instrument_id": {
            "format": "uuid",
            "title": "Instrument Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "as_of_date",
          "question",
          "horizon"
        ],
        "title": "EvidenceCaseResponse",
        "type": "object"
      },
      "EvidenceDataQualityResponse": {
        "properties": {
          "conflict_detected": {
            "title": "Conflict Detected",
            "type": "boolean"
          },
          "freshness_status": {
            "title": "Freshness Status",
            "type": "string"
          },
          "invalid_days": {
            "title": "Invalid Days",
            "type": "integer"
          },
          "observed_trading_days": {
            "title": "Observed Trading Days",
            "type": "integer"
          },
          "quality_status": {
            "title": "Quality Status",
            "type": "string"
          },
          "suspended_days": {
            "title": "Suspended Days",
            "type": "integer"
          },
          "target_trading_days": {
            "title": "Target Trading Days",
            "type": "integer"
          },
          "valid_price_days": {
            "title": "Valid Price Days",
            "type": "integer"
          }
        },
        "required": [
          "freshness_status",
          "quality_status",
          "target_trading_days",
          "observed_trading_days",
          "valid_price_days",
          "invalid_days",
          "suspended_days",
          "conflict_detected"
        ],
        "title": "EvidenceDataQualityResponse",
        "type": "object"
      },
      "EvidenceFactorResponse": {
        "properties": {
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "factor_key": {
            "title": "Factor Key",
            "type": "string"
          },
          "observed_date": {
            "format": "date",
            "title": "Observed Date",
            "type": "string"
          },
          "quality_status": {
            "title": "Quality Status",
            "type": "string"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          },
          "unit": {
            "description": "Measurement unit supplied by the frozen factor contract; interpret value using this field without percentage or currency rescaling.",
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "window": {
            "title": "Window",
            "type": "integer"
          }
        },
        "required": [
          "factor_key",
          "value",
          "unit",
          "window",
          "observed_date",
          "quality_status",
          "source_kind",
          "source_ref",
          "evidence_id"
        ],
        "title": "EvidenceFactorResponse",
        "type": "object"
      },
      "EvidenceInstrumentResponse": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "instrument_id": {
            "format": "uuid",
            "title": "Instrument Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "symbol",
          "name",
          "exchange",
          "currency"
        ],
        "title": "EvidenceInstrumentResponse",
        "type": "object"
      },
      "EvidenceMarketSnapshotResponse": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "latest_close": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Close"
          },
          "latest_trade_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Trade Date"
          },
          "observed_trading_days": {
            "title": "Observed Trading Days",
            "type": "integer"
          },
          "suspended_days": {
            "title": "Suspended Days",
            "type": "integer"
          },
          "valid_price_days": {
            "title": "Valid Price Days",
            "type": "integer"
          }
        },
        "required": [
          "latest_trade_date",
          "latest_close",
          "currency",
          "observed_trading_days",
          "valid_price_days",
          "suspended_days"
        ],
        "title": "EvidenceMarketSnapshotResponse",
        "type": "object"
      },
      "EvidencePackResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/EvidenceCaseResponse"
          },
          "data_quality": {
            "$ref": "#/components/schemas/EvidenceDataQualityResponse"
          },
          "factor_set_key": {
            "title": "Factor Set Key",
            "type": "string"
          },
          "factor_set_version": {
            "title": "Factor Set Version",
            "type": "string"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/EvidenceFactorResponse"
            },
            "title": "Factors",
            "type": "array"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "instrument": {
            "$ref": "#/components/schemas/EvidenceInstrumentResponse"
          },
          "market_snapshot": {
            "$ref": "#/components/schemas/EvidenceMarketSnapshotResponse"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Fields",
            "type": "array"
          },
          "pack_hash": {
            "title": "Pack Hash",
            "type": "string"
          },
          "pack_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Id"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSourceReferenceResponse"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "pack_id",
          "case",
          "instrument",
          "market_snapshot",
          "factors",
          "data_quality",
          "missing_fields",
          "warnings",
          "source_refs",
          "schema_version",
          "factor_set_key",
          "factor_set_version",
          "pack_hash",
          "generated_at"
        ],
        "title": "EvidencePackResponse",
        "type": "object"
      },
      "EvidenceSourceReferenceResponse": {
        "properties": {
          "observed_date": {
            "format": "date",
            "title": "Observed Date",
            "type": "string"
          },
          "quality_status": {
            "title": "Quality Status",
            "type": "string"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          }
        },
        "required": [
          "source_kind",
          "source_ref",
          "observed_date",
          "quality_status",
          "revision"
        ],
        "title": "EvidenceSourceReferenceResponse",
        "type": "object"
      },
...
      "ExternalArtifactResponse": {
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "logical_uri": {
            "title": "Logical Uri",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "artifact_id",
          "run_id",
          "logical_uri",
          "content_hash",
          "media_type",
          "size_bytes",
          "created_at"
        ],
        "title": "ExternalArtifactResponse",
        "type": "object"
      },
      "ExternalObservationResponse": {
        "properties": {
          "admission_status": {
            "title": "Admission Status",
            "type": "string"
          },
          "artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "instrument_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_uri": {
            "title": "Source Uri",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "required": [
          "observation_id",
          "run_id",
          "observed_at",
          "as_of",
          "source_uri",
          "producer",
          "payload",
          "admission_status"
        ],
        "title": "ExternalObservationResponse",
        "type": "object"
      },
      "ExternalWorkflowRunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExternalWorkflowRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "ExternalWorkflowRunListResponse",
        "type": "object"
      },
      "ExternalWorkflowRunResponse": {
        "properties": {
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "intake_status": {
            "title": "Intake Status",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "producer_status": {
            "title": "Producer Status",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "producer",
          "schema_version",
          "producer_status",
          "intake_status",
          "started_at"
        ],
        "title": "ExternalWorkflowRunResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchExternalEvidenceResponse": {
        "properties": {
          "admission": {
            "additionalProperties": true,
            "title": "Admission",
            "type": "object"
          },
          "artifact_content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Content Hash"
          },
          "artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_uri": {
            "title": "Source Uri",
            "type": "string"
          }
        },
        "required": [
          "evidence_id",
          "observation_id",
          "run_id",
          "artifact_id",
          "artifact_content_hash",
          "observed_at",
          "as_of",
          "source_uri",
          "producer",
          "payload",
          "admission",
          "content_hash"
        ],
        "title": "ResearchExternalEvidenceResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "invest-infra-v2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/external-workflows": {
      "get": {
        "operationId": "list_external_workflows_api_v1_external_workflows_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalWorkflowRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List External Workflows",
        "tags": [
          "external-workflows"
        ]
      }
    },
    "/api/v1/external-workflows/{run_id}": {
      "get": {
        "operationId": "get_external_workflow_api_v1_external_workflows__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalWorkflowRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get External Workflow",
        "tags": [
          "external-workflows"
        ]
      }
    },
    "/api/v1/external-workflows/{run_id}/artifacts": {
      "get": {
        "operationId": "list_external_artifacts_api_v1_external_workflows__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExternalArtifactResponse"
                  },
                  "title": "Response List External Artifacts Api V1 External Workflows  Run Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List External Artifacts",
        "tags": [
          "external-workflows"
        ]
      }
    },
    "/api/v1/external-workflows/{run_id}/observations": {
      "get": {
        "operationId": "list_external_observations_api_v1_external_workflows__run_id__observations_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExternalObservationResponse"
                  },
                  "title": "Response List External Observations Api V1 External Workflows  Run Id  Observations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List External Observations",
        "tags": [
          "external-workflows"
        ]
      }
    },
...
    "/api/v1/research-cases/{case_id}/evidence": {
      "get": {
        "operationId": "get_research_case_evidence_api_v1_research_cases__case_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidencePackResponse"
                  },
                  "title": "Response Get Research Case Evidence Api V1 Research Cases  Case Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research Case Evidence",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/research-cases/{case_id}/external-observations/{observation_id}/evidence": {
      "post": {
        "operationId": "link_external_observation_api_v1_research_cases__case_id__external_observations__observation_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchExternalEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link External Observation",
        "tags": [
          "research"
        ]
      }
    },
...
  }
}